adds chaincode_labs_exercise.py test
Adds a new test named chaincode_labs_exercise.py. The test creates two nodes that aren't connected, mines a block on node 0, connects the nodes and syncs the new block, and asserts the nodes best block now matches. The purpose of this test is purely an exercise to get accepted to the chaincode labs seminar.